[fix] ディレクトリの修正
