--- src/internal/config/config.yaml
@@ -1,21 +1,0 @@
-db:
-  driver: sqlite3
-  path: ./internal/db/sqlite3/are_you_free.db
-  file: ./internal/db/sqlite3/are_you_free.sql
-
-filepath:
-  CSS: internal/web/static/css
-  JS: internal/web/static/js
-  image: internal/web/img
-  icon: internal/web/icon
-
-session:
-  name: someSessionName
-  keyName: UserID
-  secretKey: someSessionKey
-  Path: /
-  Domain: ""
-  MaxAgeSec: 86400
-  MaxAgeDay: 7
-  Secure: false
-  httpOnly: true
