+++ src/internal/handlers/templates/templates.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"fmt"
 	"net/http"
-	"src/internal/config"
 	"src/internal/entity"
+	"src/internal/infra/config"
 	"src/internal/utils/strings"
 	"src/internal/utils/times"
 	"strconv"
